refactor: split docs markdown pure renderer
